+++ sql/registrar_certification_ddl.sql
@@ -11,7 +11,9 @@ CREATE TABLE registrar_certification
     valid_from date NOT NULL, --  registrar certification valid from
     valid_until date, --  registrar certification valid until
     classification classification_type NOT NULL, -- registrar certification result checked 0-5
-    eval_file_id integer NOT NULL CONSTRAINT registrar_certification_eval_file_id_fkey REFERENCES files(id) -- link to pdf file
+    eval_file_id integer NOT NULL CONSTRAINT registrar_certification_eval_file_id_fkey REFERENCES files(id), -- link to pdf file
+    uuid UUID NOT NULL UNIQUE DEFAULT gen_random_uuid(), -- external reference
+    eval_file_uuid UUID NOT NULL -- reference to external pdf file
 );
 
 -- check whether registrar_certification life is valid
@@ -59,6 +61,56 @@ CREATE TRIGGER "trigger_registrar_certification"
   AFTER INSERT OR UPDATE ON registrar_certification
   FOR EACH ROW EXECUTE PROCEDURE registrar_certification_life_check();
 
+-- synchronize eval_file_id with eval_file_uuid through files table
+CREATE OR REPLACE FUNCTION registrar_certification_eval_file_id_sync()
+RETURNS "trigger" AS $$
+BEGIN
+    IF TG_OP = 'INSERT' THEN
+        -- old clients insert id
+        IF NEW.eval_file_id IS NOT NULL THEN
+            SELECT f.uuid
+              FROM files f
+             WHERE f.id = NEW.eval_file_id
+              INTO NEW.eval_file_uuid;
+            RETURN NEW;
+        END IF;
+        -- new clients insert uuid
+        IF NEW.eval_file_uuid IS NOT NULL THEN
+            SELECT f.id
+              FROM files f
+             WHERE f.uuid = NEW.eval_file_uuid
+              INTO NEW.eval_file_id;
+            RETURN NEW;
+        END IF;
+    ELSEIF TG_OP = 'UPDATE' THEN
+        -- old clients update id
+        IF NEW.eval_file_id IS NOT NULL AND NEW.eval_file_id <> OLD.eval_file_id THEN
+            SELECT f.uuid
+              FROM files f
+             WHERE f.id = NEW.eval_file_id
+              INTO NEW.eval_file_uuid;
+            RETURN NEW;
+        END IF;
+        -- new clients update uuid
+        IF NEW.eval_file_uuid IS NOT NULL AND NEW.eval_file_uuid <> OLD.eval_file_uuid THEN
+            SELECT f.id
+              FROM files f
+             WHERE f.uuid = NEW.eval_file_uuid
+              INTO NEW.eval_file_id;
+            RETURN NEW;
+        END IF;
+    END IF;
+    RETURN NEW;
+END
+$$ LANGUAGE plpgsql;
+
+COMMENT ON FUNCTION registrar_certification_eval_file_id_sync()
+    IS 'synchronize `eval_file_id` with `eval_file_uuid` through `files` table';
+
+CREATE TRIGGER "trigger_registrar_certification_eval_file_id_sync"
+    BEFORE INSERT OR UPDATE ON registrar_certification
+    FOR EACH ROW EXECUTE PROCEDURE registrar_certification_eval_file_id_sync();
+
 
 CREATE INDEX registrar_certification_valid_from_idx ON registrar_certification(valid_from);
 CREATE INDEX registrar_certification_valid_until_idx ON registrar_certification(valid_until);
@@ -73,6 +125,9 @@ COMMENT ON COLUMN registrar_certification.classification IS
     'registrar certification result checked 0-5';
 COMMENT ON COLUMN registrar_certification.eval_file_id IS
     'evaluation pdf file link';
+COMMENT ON COLUMN registrar_certification.uuid IS 'uuid for external reference';
+COMMENT ON COLUMN registrar_certification.eval_file_uuid IS
+    'evaluation pdf file external link';
 
 CREATE TABLE registrar_group
 (